telegram-notification-bot/pyproject.toml
[project]
name = "telegram-notification-bot"
version = "0.1.0"
description = ""
authors = [
    {name = "Ana Luiza",email = "106170960+analuizataveira@users.noreply.github.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "python-telegram-bot (>=21.10,<22.0)",
    "python-dotenv (>=1.0.1,<2.0.0)"
]

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
pyinstaller = "^6.12.0"


[tool.poetry.group.dev.dependencies]
pyinstaller = "^6.12.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
